+++ libft/srcs/ft_flag_c_maj.c
@@ -6,7 +6,7 @@
 /*   By: ldedier <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/05 21:37:22 by ldedier           #+#    #+#             */
-/*   Updated: 2018/05/31 04:02:22 by ldedier          ###   ########.fr       */
+/*   Updated: 2018/12/08 18:10:16 by emuckens         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,7 +53,7 @@ int		ft_pf_c_maj(t_pf *pf, va_list va, char c)
 	(void)c;
 	pf->var.integer = va_arg(va, int);
 	if ((len = ft_wchar_len(pf->var.integer)) == -1 ||
-			((unsigned int)len > MB_CUR_MAX && pf->var.integer >= 255))
+			(len > MB_CUR_MAX && pf->var.integer >= 255))
 		return (-1);
 	ft_add_padding_pre(pf, len);
 	ft_put_wchar(pf->var.integer, pf);
